reorganize sm90 tests and benches
