+++ tests/test_kda_sm100_chunk_vs_naive.py
