+++ tests/test_kda_sm100_intracard_cp.py
@@ -46,9 +46,9 @@ BT, K, V = 64, 128, 128
 DEVICE = "cuda"
 # Tolerances aligned with existing cuLA tests:
 #   * Same-kernel (CP scheduling only): torch.testing.assert_close(atol=1e-2, rtol=1e-2)
-#                                       — matches tests/test_chunk_delta_h.py CP block
+#                                       — matches tests/test_kda_sm100_recurrence.py CP block
 #   * Cross-impl / vs ref:              fla.utils.assert_close(ratio=...)
-#                                       — matches tests/test_kda_compare_fla.py
+#                                       — matches tests/test_kda_sm100_chunk_vs_fla.py
 ATOL_SAME_KERNEL = 1e-2
 RTOL_SAME_KERNEL = 1e-2
 RATIO_VS_REF = 0.005  # RMSE / RMS(ref) — matches FLA test_gated_delta.py fwd
@@ -223,7 +223,7 @@ def pytorch_ref(k, w, u, *, gk=None, initial_state=None, cu_seqlens, save_new_va
 
 
 def _assert_same_kernel(name, actual, ref):
-    """torch.testing.assert_close — matches tests/test_chunk_delta_h.py."""
+    """torch.testing.assert_close — matches tests/test_kda_sm100_recurrence.py."""
     if actual is None or ref is None:
         assert actual is ref, f"{name}: one is None and other isn't"
         return
@@ -289,7 +289,7 @@ def test_cp_autodispatch_matches_baseline(seq_lens, H, use_gk):
     """CP auto-dispatch output equals no-CP baseline (same kernel).
 
     Tolerance: `torch.testing.assert_close(atol=1e-2, rtol=1e-2)` —
-    matches the CP block in tests/test_chunk_delta_h.py.
+    matches the CP block in tests/test_kda_sm100_recurrence.py.
     """
     k, w, u, gk, _, cu = make_varlen_inputs(seq_lens, H, use_gk=use_gk)
     h_base, v_base, _ = run_cula_no_cp(k, w, u, gk, None, cu)
