+++ tests/test_kda_sm100_recurrence.py
